test(bare-metal): make offline qualification reproducible
Plain-language summary:

- Refresh changed same-version internal wheels without re-resolving package versions.
- Place complete-scenario orchestration coverage at the integration-test level.
- Preserve every existing adverse-path and parsing assertion.

Architectural summary:

Locked package setup now separates dependency selection from artifact refresh:
checked-in lockfiles select versions, while explicit reinstall inventories load
rebuilt same-version wheels. The bare-metal storefront distribution target adds
the three corrected registry-client, config and Alkahest producers without
claiming complete closure over every reinstalled internal package. Offline tests
that execute the real complete-deal scenario are classified as integration
coverage, while direct envelope and access-proof parsing remains unit coverage.

Change groups:

- Add registry-client, config and Alkahest prerequisites to the bare-metal
  storefront distribution target.
- Explicitly reinstall registry-client and config in the storefront environment.
- Use locked sync for bare-metal storefront, Alkahest and E2E setup.
- Remove upgrade-driven version re-resolution while preserving exact reinstall
  inventories of 15 storefront, 2 Alkahest and 24 E2E packages.
- Add focused Makefile regressions for producer prerequisites, locked setup and
  exact reinstall inventories.
- Move 15 complete-scenario orchestration tests to the existing integration
  layer and retain the 3 direct parsing test functions in the unit layer.
- Record deterministic packaging and test ownership in the active change.

Boundary decisions and deferrals:

- The root storefront target builds the three corrected producers; other
  reinstalled wheels remain prerequisites of the established build flow.
- No complete producer-closure claim or broader build-graph expansion is made.
- No dependency versions, lockfiles, package metadata or platform markers
  change.
- The authoritative production complete-deal scenario remains unchanged.
- Failure-after-purchase cleanup, settlement ordering, no-op rejection, stale
  leased listings, SSH pinning, teardown failure, trusted discovery and run-log
  assertions remain intact.
- Actual-host setup and rollback remain operator-controlled and separately
  gated.
- No production Python, infrastructure, deployment, funding, SSH, vLLM, cache
  or GPU behavior changes.
- Final offline and live qualification tasks remain open.

Validation:

- Parent canonical locked bare-metal storefront setup and tests passed.
- Parent canonical locked Alkahest setup and tests passed.
- Parent canonical locked E2E setup and tests passed.
- Parent confirmed all three lockfiles remained unchanged.
- Parent core storefront suite: 161 passed.
- Independent AST review accounted for all 18 original test functions: 15
  integration functions remained identical and 3 parsing functions remained
  in unit coverage.
- Independent package regressions: 3 passed.
- Independent relocated unit and integration scenario suites: 28 passed.
- Focused local package regressions: 3 passed.
- Focused local relocated scenario suites: 28 passed.
- Ruff passed for the changed Python tests using the existing environment.
- git diff --check passed.
- make check-comment-hygiene passed.
- Strict OpenSpec validation was unavailable because the CLI is not installed.
- No live qualification was run or claimed.

Result:

Deterministic locked wheel refresh and test placement are verified offline and
ready for commit review. The overall bare-metal goal remains incomplete:
remaining qualification cleanup and any live validation require separate gates.
