+++ e2e-tests/tests/integration/__init__.py
@@ -1,0 +1,1 @@
+"""Offline integration coverage for E2E scenario orchestration."""
