+++ kit/alkahest/Makefile
@@ -6,9 +6,9 @@ build: ## Build the wheel into the shared .dist/
 	uv build --wheel --out-dir $(DIST_DIR)
 
 reinit: ## Refresh changed internal settlement dependencies.
-	uv sync --dev --find-links $(DIST_DIR) \
-		--upgrade-package arkhai-core --reinstall-package arkhai-core \
-		--upgrade-package arkhai-kit-settlement-runtime --reinstall-package arkhai-kit-settlement-runtime
+	uv sync --locked --dev --find-links $(DIST_DIR) \
+		--reinstall-package arkhai-core \
+		--reinstall-package arkhai-kit-settlement-runtime
 
 test: test-unit ## Run the default test suite
